requirements.txt
mcp>=0.1.0
azure-ai-projects>=1.0.0
azure-identity>=1.12.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0